Penambahan Comment Controller dll

+++ models/Barang.go
@@ -6,7 +6,7 @@ type UserBarang struct {
 	Desc   string `json:desc`
 	Image  string `json:image`
 	Star   string `json:star`
-	Price  int    `json:price`
+	Price  int64  `json:price`
 	UserID string `json:userid`
 	User   User   `json:"user";gorm:"foreignkey:UserID"`
 }
